requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
google-genai>=1.73.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0
sentence-transformers>=2.2.0
python-dotenv>=0.19.0
